Replaced set with list to collect all the symbols for prefetching.
This is a sub-step before second step of the refactoring for #17.

This change is made possible by the previous refactoring #16, which used memoization for the stock symbols.
Before that refactoring, it was required to use a set to collect all the symbols without repetition.
As the memoization requires the first appearance of the symbol to be handled differently, that is,
memoizing the symbol, we can piggyback this change.